Move JobApplicationNote entity to the separate module
