+++ src/database.config.ts
@@ -8,7 +8,7 @@ import { UserCodeVerification } from './modules/users/entities/user.code.verific
 import { Contact } from './modules/contacts/entities/contact.entity';
 import { ContactEmail } from './modules/contacts/entities/contact-emails.entity';
 import { ContactPhone } from './modules/contacts/entities/contact-phones.entity';
-import { JobApplicationNote } from './modules/job-applications/entities/job-application-note.entity';
+import { JobApplicationNote } from './modules/job-application-notes/entities/job-application-note.entity';
 
 if (process.env.NODE_ENV !== 'test') {
   config();
