+++ src/modules/job-applications/entities/job-application.entity.ts
@@ -2,7 +2,7 @@ import { BaseEntity } from '../../../entities/base.entity';
 import { BoardColumn } from '../../board-columns/entities/board-column.entity';
 import { Column, Entity, JoinColumn, ManyToMany, ManyToOne, OneToMany } from 'typeorm';
 import { Contact } from '../../contacts/entities/contact.entity';
-import { JobApplicationNote } from './job-application-note.entity';
+import { JobApplicationNote } from '../../job-application-notes/entities/job-application-note.entity';
 
 @Entity('job_applications')
 export class JobApplication extends BaseEntity {
