+++ src/modules/job-applications/job-applications.module.ts
@@ -6,7 +6,7 @@ import { JobApplication } from './entities/job-application.entity';
 import { BoardColumn } from '../board-columns/entities/board-column.entity';
 import { JobApplicationMapper } from './job-applications.mapper';
 import { Board } from '../boards/entities/board.entity';
-import { JobApplicationNote } from './entities/job-application-note.entity';
+import { JobApplicationNote } from '../job-application-notes/entities/job-application-note.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([JobApplication, JobApplicationNote, BoardColumn, Board])],
